applications/user/requirements.txt
phonenumbers
django-phonenumber-field